+++ api/src/routes/controllers/userType.js
@@ -13,18 +13,61 @@ const { SECRET } = process.env;
 
 
 const signIn = async (req, res) => {
+
+    const { name, idUser, gmail, photograph, userType } = req.body;
+
     try{
-        const { email, userType } = req.body;
-        console.log(email, userType)
+        
         if(userType === 'junior'){
             
 
-            const user = await Juniors.findOne({gmail: email})
+            const user = await Juniors.findOne({gmail: gmail})
             if(!user){
-                return res.status(404).json({auth: false, message: 'email no registrado'})
+                
+                const juniorsCreate = await Juniors.create({
+                    _id: idUser,
+                    name: name,
+                    gmail:gmail,
+                    photograph: photograph || 'https://www.w3schools.com/howto/img_avatar.png',
+                })
+
+                const token = jwt.sign({id: juniorsCreate._id}, SECRET, {
+                    expiresIn: 60 * 60 * 24
+                })
+                
+                res.json({auth: true, token: token, user: juniorsCreate});
+                return
             }
 
-            const token = jwt.sign({id: user._id}, SECRET, {
+            const token = jwt.sign({id: idUser}, SECRET, {
+                expiresIn: 60 * 60 * 24
+            })
+
+            res.json({auth: true, token: token, user: user})
+        }
+
+        if(userType === 'company'){
+            
+
+            const user = await Company.findOne({gmail: gmail})
+            if(!user){
+                
+                const CompanyCreate = await Company.create({
+                    _id: idUser,
+                    name: name,
+                    gmail:gmail,
+                    photograph: photograph || 'https://www.w3schools.com/howto/img_avatar.png',
+                })
+
+                const token = jwt.sign({id: CompanyCreate._id}, SECRET, {
+                    expiresIn: 60 * 60 * 24
+                })
+                
+                res.json({auth: true, token: token, user: CompanyCreate});
+                return
+            }
+
+            const token = jwt.sign({id: idUser}, SECRET, {
                 expiresIn: 60 * 60 * 24
             })
